Move decay_parameter into its own header
