+++ agency/bulk_invoke.hpp
@@ -4,6 +4,7 @@
 #include <agency/executor_traits.hpp>
 #include <agency/execution_agent.hpp>
 #include <agency/functional.hpp>
+#include <agency/detail/bulk_functions/decay_parameter.hpp>
 #include <agency/detail/bulk_functions/shared_parameter.hpp>
 #include <agency/detail/bulk_functions/bind_agent_local_parameters.hpp>
 #include <agency/detail/is_call_possible.hpp>
@@ -22,42 +23,6 @@ namespace agency
 {
 namespace detail
 {
-
-
-// this metafunction computes the type of the parameter passed to a user function
-// given then type of parameter passed to bulk_invoke/bulk_async/etc.
-template<class T>
-struct decay_parameter
-{
-  template<class U>
-  struct lazy_add_lvalue_reference
-  {
-    using type = typename std::add_lvalue_reference<typename U::type>::type;
-  };
-
-  // first decay the parameter
-  using decayed_type = typename std::decay<T>::type;
-
-  // when passing a parameter to the user's function:
-  // if the parameter is a future, then we pass a reference to its value type
-  // otherwise, we pass a copy of the decayed_type
-  using type = typename detail::lazy_conditional<
-    is_future<decayed_type>::value,
-    lazy_add_lvalue_reference<future_value<decayed_type>>,
-    identity<decayed_type>
-  >::type;
-};
-
-template<class T>
-using decay_parameter_t = typename decay_parameter<T>::type;
-
-
-template<size_t level, class T, class... Args>
-struct decay_parameter<shared_parameter<level,T,Args...>>
-{
-  // shared_parameters are passed to the user function by reference
-  using type = T&;
-};
 
 
 template<class ExecutorTraits, class AgentTraits, class Function, size_t... UserArgIndices>
